bugfix: RuntimeError: expected scalar type Long but found Float

+++ src/models/make_lightning.py
@@ -54,7 +54,7 @@ class SegSubLightning(pl.LightningModule):
     def training_step(self, batch):
         item, inputs = batch
         outputs = self.forward(inputs)
-        loss = self.criterion(outputs, inputs['labels'].float())
+        loss = self.criterion(outputs, inputs['labels'])
         self.log('train/loss', loss, on_epoch=True, sync_dist=True)
 
         return loss
@@ -62,7 +62,7 @@ class SegSubLightning(pl.LightningModule):
     def validation_step(self, batch, batch_idx):
         item, inputs = batch
         outputs = self.forward(inputs)
-        loss = self.criterion(outputs, inputs['labels'].float())
+        loss = self.criterion(outputs, inputs['labels'])
         self.log('val/loss', loss, on_epoch=True, sync_dist=True)
         outputs = (F.sigmoid(outputs) > 0.5).type(torch.uint8)
         self.metrics.update(outputs, inputs['labels'])
